update and add to card in detailed page
